# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trasteros API",
        "description": "API para la gesti\u00f3n de trasteros, contratos, clientes y finanzas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/contratos/trastero/{trasteroId}": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Contratos por trastero",
                "description": "Obtiene todos los contratos de un trastero especifico",
                "operationId": "get_contratos_by_trastero",
                "parameters": [
                    {
                        "name": "trasteroId",
                        "in": "path",
                        "description": "ID del trastero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "onlyActivos",
                        "in": "query",
                        "description": "Solo contratos activos",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contratos del trastero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contrato"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ingresos/trastero/{trasteroId}": {
            "get": {
                "tags": [
                    "Ingresos"
                ],
                "summary": "Ingresos por trastero",
                "description": "Obtiene todos los ingresos de un trastero especifico",
                "operationId": "get_ingresos_by_trastero",
                "parameters": [
                    {
                        "name": "trasteroId",
                        "in": "path",
                        "description": "ID del trastero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ingresos del trastero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ingreso"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/trasteros": {
            "get": {
                "tags": [
                    "Trasteros"
                ],
                "summary": "Listar trasteros",
                "description": "Obtiene la lista de todos los trasteros",
                "operationId": "get_trasteros_list",
                "parameters": [
                    {
                        "name": "localId",
                        "in": "query",
                        "description": "Filtrar por local",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "disponible",
                                "ocupado",
                                "reservado",
                                "mantenimiento"
                            ]
                        }
                    },
                    {
                        "name": "onlyActive",
                        "in": "query",
                        "description": "Solo activos",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de trasteros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Trastero"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trasteros"
                ],
                "summary": "Crear trastero",
                "description": "Crea un nuevo trastero",
                "operationId": "post_trasteros_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "localId",
                                    "numero",
                                    "superficie",
                                    "precioMensual"
                                ],
                                "properties": {
                                    "localId": {
                                        "type": "integer"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "A-01"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "superficie": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    },
                                    "precioMensual": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "estado": {
                                        "type": "string",
                                        "default": "disponible",
                                        "enum": [
                                            "disponible",
                                            "ocupado",
                                            "reservado",
                                            "mantenimiento"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trastero creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trastero"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validaci\u00f3n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Trastero duplicado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/trasteros/local/{localId}": {
            "get": {
                "tags": [
                    "Trasteros"
                ],
                "summary": "Trasteros por local",
                "description": "Obtiene todos los trasteros de un local espec\u00edfico",
                "operationId": "get_trasteros_by_local",
                "parameters": [
                    {
                        "name": "localId",
                        "in": "path",
                        "description": "ID del local",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de trasteros del local",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Trastero"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/trasteros/disponibles": {
            "get": {
                "tags": [
                    "Trasteros"
                ],
                "summary": "Trasteros disponibles",
                "description": "Obtiene trasteros disponibles. Con parametro fecha, devuelve los que estaran disponibles en esa fecha.",
                "operationId": "get_trasteros_disponibles",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "description": "Fecha para verificar disponibilidad (Y-m-d, default: hoy)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de trasteros disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Trastero"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginatedMeta": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "Contrato": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "trasteroId": {
                        "type": "integer"
                    },
                    "clienteId": {
                        "type": "integer"
                    },
                    "fechaInicio": {
                        "type": "string",
                        "format": "date"
                    },
                    "fechaFin": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "precioMensual": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "fianza": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "fianzaPagada": {
                        "type": "boolean"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "activo",
                            "pendiente",
                            "finalizado",
                            "cancelado"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Ingreso": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "contratoId": {
                        "type": "integer"
                    },
                    "concepto": {
                        "type": "string",
                        "example": "Mensualidad Enero 2024"
                    },
                    "importe": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "fechaPago": {
                        "type": "string",
                        "format": "date"
                    },
                    "categoria": {
                        "type": "string",
                        "enum": [
                            "alquiler",
                            "fianza",
                            "otros"
                        ]
                    },
                    "metodoPago": {
                        "type": "string",
                        "enum": [
                            "efectivo",
                            "transferencia",
                            "tarjeta",
                            "domiciliacion",
                            "bizum"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "JWT Authorization header usando el esquema Bearer. Ejemplo: \"Authorization: Bearer {token}\"",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}